bugfix(DPAV-1151): Redirect to landing page on 401/403
The previous fix wasn't working as we are unable to call any apis
once the authentication is expired, they will all result in the
oauth2-proxy 403 HTML response.

Now we will fetch the landing page url at startup and save it for
later to use for redirecting. As they won't be able to load the
app without valid authentication this will work as intended (without
a valid cookie they would be redirected before the request made it
to the web app).